AMB-962 rename test transformer namespace to converter
