+++ docker/Transformation-Engine/src/test/java/nhsd/fhir/converter/service/converter/MedicationRequestConverterTest.java
@@ -1,7 +1,6 @@
-package nhsd.fhir.converter.service.transformers;
+package nhsd.fhir.converter.service.converter;
 
 import ca.uhn.fhir.context.FhirVersionEnum;
-import nhsd.fhir.converter.service.converter.MedicationRequestConverter;
 import org.apache.commons.io.FileUtils;
 import org.json.JSONException;
 import org.json.JSONObject;
